Fix type casting in FNV-1 and FNV-1a hash update methods to use uint8_t

+++ include/hash23/fnv_1.h
@@ -38,7 +38,7 @@ namespace hash23 {
         constexpr void update(T const &data) {
             for (auto const &byte : data) {
                 hash_ *= prime();
-                hash_ ^= static_cast<std::size_t>(byte);
+                hash_ ^= static_cast<std::uint8_t>(byte);
             }
         }
