+++ include/hash23/fnv_1a.h
@@ -37,7 +37,7 @@ namespace hash23 {
             requires std::ranges::contiguous_range<T>
         constexpr void update(T const &data) {
             for (auto const &byte: data) {
-                hash_ ^= static_cast<std::size_t>(byte);
+                hash_ ^= static_cast<std::uint8_t>(byte);
                 hash_ *= prime();
             }
         }
